+++ PgRoutiner/Builder/Dump/PgDumpVersion.cs
@@ -2,55 +2,55 @@
 
 public static class PgDumpVersion
 {
-    public static bool Check(PgDumpBuilder builder)
+    public static bool Check(PgDumpBuilder builder, bool restore = false)
     {
-        var option = (typeof(Settings).GetField($"{builder.PgDumpName}Args").GetValue(null) as Arg).Original;
+        var option = (typeof(Settings).GetField(name: $"{(restore ? nameof(Settings.PgRestore) : nameof(Settings.PgDump))}Args").GetValue(null) as Arg).Original;
         var connVersion = builder.Connection.ServerVersion.Split(".").First();
         string fullDumpVersion;
         string dumpVersion;
         try
         {
-            fullDumpVersion = builder.GetDumpVersion();
+            fullDumpVersion = builder.GetDumpVersion(restore);
             dumpVersion = fullDumpVersion.Split(".").First();
         }
         catch (Exception e)
         {
-            PgDumpError(connVersion, e, option);
+            PgDumpError(connVersion, e, option, restore: restore);
             return false;
         }
 
         if (!string.Equals(connVersion, dumpVersion))
         {
-            builder.SetPgDumpName(string.Format(Settings.Value.GetPgDumpFallback(), connVersion));
+            builder.SetPgDumpName(string.Format(restore ? Settings.Value.GetPgRestoreFallback() : Settings.Value.GetPgDumpFallback(), connVersion));
             try
             {
-                fullDumpVersion = builder.GetDumpVersion();
+                fullDumpVersion = builder.GetDumpVersion(restore);
                 dumpVersion = fullDumpVersion.Split(".").First();
             }
             catch (Exception e)
             {
-                PgDumpError(connVersion, e, option);
+                PgDumpError(connVersion, e, option, restore: restore);
                 return false;
             }
 
             var value = typeof(Settings).GetProperty(builder.PgDumpName).GetValue(Settings.Value);
             if (!string.Equals(connVersion, dumpVersion))
             {
-                PgDumpMistmatch(builder.Connection, connVersion, fullDumpVersion, dumpVersion, value, option);
+                PgDumpMistmatch(builder.Connection, connVersion, fullDumpVersion, dumpVersion, value, option, restore: restore);
                 return false;
             }
             Program.WriteLine(ConsoleColor.Yellow, "",
-                $"WARNING: Using fall-back path for pg_dump: {value}. To remove this warning set the {option} setting to point to this path.",
+                $"WARNING: Using fall-back path for {(restore ? "pg_restore" : "pg_dump")}: {value}. To remove this warning set the {option} setting to point to this path.",
                 "");
         }
         return true;
     }
 
     private static void PgDumpMistmatch(NpgsqlConnection connection, string connVersion, string fullDumpVersion, string dumpVersion,
-        object value, string option)
+        object value, string option, bool restore)
     {
         Program.WriteLine("");
-        Program.WriteLine(ConsoleColor.Red, "ERROR: It looks like pg_dump version mismatch: ");
+        Program.WriteLine(ConsoleColor.Red, $"ERROR: It looks like {(restore ? "pg_restore" : "pg_dump")} version mismatch: ");
 
         Program.Write(ConsoleColor.Red, "- Connection ");
         Program.Write(ConsoleColor.Red, Settings.Value.Connection);
@@ -67,7 +67,7 @@ public static class PgDumpVersion
         Program.WriteLine(ConsoleColor.Red, ")");
 
         Program.Write(ConsoleColor.Red, "- Default ");
-        Program.Write(ConsoleColor.Red, "pg_dump");
+        Program.Write(ConsoleColor.Red, $"{(restore ? "pg_restore" : "pg_dump")}");
         Program.Write(ConsoleColor.Red, " path (");
         Program.Write(ConsoleColor.Red, $"{value}");
         Program.Write(ConsoleColor.Red, ") is version ");
@@ -79,7 +79,7 @@ public static class PgDumpVersion
         Program.WriteLine(ConsoleColor.Red,
             "You will not be able to create database dump files (schema file, data dump files, object tree dir or schema diff), unless versions match!");
         Program.Write(ConsoleColor.Red, "Make sure you have access to ");
-        Program.Write(ConsoleColor.Red, "pg_dump");
+        Program.Write(ConsoleColor.Red, $"{(restore ? "pg_restore" : "pg_dump")}");
         Program.Write(ConsoleColor.Red, " version ");
         Program.Write(ConsoleColor.Red, connVersion);
         Program.Write(ConsoleColor.Red, " and, the settings key ");
@@ -87,17 +87,17 @@ public static class PgDumpVersion
         Program.WriteLine(ConsoleColor.Red, " points to the correct path.");
     }
 
-    private static void PgDumpError(string connVersion, Exception e, string option)
+    private static void PgDumpError(string connVersion, Exception e, string option, bool restore)
     {
         Program.WriteLine("");
-        Program.Write(ConsoleColor.Red, "ERROR - pg_dump returned an error: ");
+        Program.Write(ConsoleColor.Red, $"ERROR - {(restore ? "pg_restore" : "pg_dump")} returned an error: ");
         Program.WriteLine(ConsoleColor.Red, e.Message);
 
         Program.WriteLine(ConsoleColor.Red, "PostgreSQL might not be installed on your computer.");
         Program.WriteLine(ConsoleColor.Red,
             "You will not be able to create database dump files (schema file, data dump file or object tree dir), unless versions match!");
         Program.Write(ConsoleColor.Red, "Make sure you have access to ");
-        Program.Write(ConsoleColor.Red, "pg_dump");
+        Program.Write(ConsoleColor.Red, $"{(restore ? "pg_restore" : "pg_dump")}");
         Program.Write(ConsoleColor.Red, " version ");
         Program.Write(ConsoleColor.Red, connVersion);
         Program.Write(ConsoleColor.Red, " and, the settings key ");